refactor: create `Header` component
